pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ffxi-strings"
version = "1.0.0"
requires-python = ">=3.11"

dependencies = [
    "loguru>=0.7",
    "pyyaml>=6.0",
]

[tool.setuptools]
packages = ["parser", "models"]
